+++ module/Soutenance/src/Soutenance/Provider/Template/PdfTemplates.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Soutenance\Provider\Template;
+
+class PdfTemplates {
+
+    const SERMENT_DU_DOCTEUR = "SERMENT_DU_DOCTEUR";
+
+}